% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_backend.R
\name{util_backend}
\alias{util_backend}
\title{Set foreach Backend}
\usage{
util_backend(par = FALSE, ncores = NULL, backend = "doParallel", sock = TRUE)
}
\arguments{
\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{backend}{Character string.
\code{foreach} backend.
The following are currently available:
\code{"doParallel"}, and \code{"doMC"}.
If the current operating system
is Windows,
and \code{backend = "doMC"}
backend is changed to \code{"doParallel"}.}

\item{sock}{Logical.
If \code{backend = "doParallel"},
you have two parallel options.
If \code{TRUE}, uses sockets.
If \code{FALSE}, uses forking.
If the current operating system
is Windows,
\code{sock = TRUE} is enforced.}
}
\description{
Set foreach Backend
}
\author{
Ivan Jacob Agaloos Pesigan
}
