% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bash.R
\name{util_bash}
\alias{util_bash}
\title{Setup Bash dotfiles}
\usage{
util_bash(overwrite = FALSE, vars = NULL)
}
\arguments{
\item{overwrite}{Logical.
Overwrite existing \code{bash} dot files in the home directory.}

\item{vars}{Named character vector.
Variables to export.
(e.g., \code{vars = c(GITHUB_PAT = "token_here", TRAVIS_TOKEN = "token_here")}.}
}
\description{
Creates the following files:
\code{.bashrc},
\code{.bash_aliases},
\code{.bash_profile},
\code{.bash_logout}
in the home directory.
}
\author{
Ivan Jacob Agaloos Pesigan
}
