% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bind.R
\name{util_bind}
\alias{util_bind}
\title{Bind External Files.}
\usage{
util_bind(
  dir = getwd(),
  format = "csv",
  pattern = "^filename.*",
  fn_column = TRUE,
  fn_column_full_path = FALSE,
  save = FALSE,
  fn = NULL,
  save_dir = getwd(),
  par = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{dir}{Character string.
Directory which contains target files.}

\item{format}{Character string.
\code{"csv"} for comma separated files.
\code{"xls"} or \verb{"xlsx} for \code{Excel} files.}

\item{pattern}{Character string.
Regular expression.
Pattern of file names.
\code{format} is appended as an extension.
For example, if \code{pattern = "^filename.*"},
and \code{format = "csv"},
the pattern used to load files will be
\code{"^filename.*\\\\.csv$"}.}

\item{fn_column}{Logical.
Save file name of source data file as a new column.}

\item{fn_column_full_path}{Logical.
Save the full path of source data file as a new column.}

\item{save}{Logical.
Save concatenated files in \code{csv} format.}

\item{fn}{Character string.
Filename to use when \code{save = TRUE}.
Uses the basename of \code{dir} if not provided.}

\item{save_dir}{Character string.
Directory used for \code{fn}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{...}{Optional arguments to pass to
\code{\link[readxl:read_excel]{readxl::read_excel()}} when \code{format \%in\% c("xls", "xlsx")},
\code{\link[=read.csv]{read.csv()}} when \code{format = "csv"},
\code{\link[=load]{load()}} when \code{format \%in\% c("Rda", "RDA", "RData", "Rdata", "rdata", "RDATA")}, or
\code{\link[=readRDS]{readRDS()}} when \code{format \%in\% c("Rds", "rds", "RDS")}.}
}
\description{
Binds \code{csv}, \code{Excel}, \code{RData}, or \code{Rds} files
in a given directory by row (\code{rbind}).
}
\details{
The function requires that files have the data structure.
For \code{R} object files.
only vectors, matrices or data frames
are allowed.
By default, \code{stringsAsFactors = FALSE}
when \code{format = "csv"}.
}
\examples{
\dontrun{
util_bind(
  dir = getwd(),
  format = "csv",
  pattern = "^filename.*",
  fn_column = TRUE,
  save = TRUE,
  fn = NULL,
  save_dir = getwd(),
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
