% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cat_sys.R
\name{util_cat_sys}
\alias{util_cat_sys}
\title{Cat System Output in knitr.}
\usage{
util_cat_sys(command, sep = "\\n")
}
\arguments{
\item{command}{the system command to be invoked, as a character string.}

\item{sep}{a character vector of strings to append after each element.}
}
\description{
Utility to \code{\link[=cat]{cat()}} \code{\link[=system]{system()}} output.
Based on https://github.com/yihui/knitr/issues/1203.
}
\examples{
util_cat_sys("date")
}
