% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_file_seq.R
\name{util_check_file_seq}
\alias{util_check_file_seq}
\title{Check Files in a Sequence}
\usage{
util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = TRUE
)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{start}{Integer.
Start of the sequence.}

\item{end}{Integer.
End of the sequence.}

\item{digits}{Integer.
Digits used in number sequence.}

\item{fn}{Character string.
Filename EXCLUDING
the sequence number,
the separator (\code{sep}), and
the file extension (\code{ext}).
The sequence number,
separator and
file extension are declared
in the other arguments.}

\item{ext}{Character string.
File extension,
for example \code{ext = "csv"}
for Comma Separated Values.}

\item{sep}{Character string.
String that separates number sequence and \code{fn}.}

\item{prefix}{Logical.
If \code{TRUE}, number sequence are at the beginning of \code{fn}.
If \code{FALSE}, number sequence are at the end of \code{fn}.}
}
\description{
Check Files in a Sequence
}
\examples{
\dontrun{
# This will search for
# {5 digits}_filename.csv
util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = TRUE
)
# This will search for
# filename_{5 digits}.csv
util_check_file_seq(
  dir = getwd(),
  start = 1L,
  end = 10L,
  digits = 5L,
  fn = "filename",
  ext = "csv",
  sep = "_",
  prefix = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
