% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_file_type.R
\name{util_check_file_type}
\alias{util_check_file_type}
\title{Check Files.}
\usage{
util_check_file_type(
  dir = getwd(),
  fn,
  file_type = "PDF document",
  remove_files = FALSE,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory which contains files.}

\item{fn}{Character vector.
Filenames.}

\item{file_type}{Character string.
File type.
For example, "PDF document" for PDF and
"EPUB document" for EPUB.}

\item{remove_files}{Logical.
Remove files that do not match the specified file type.
BE CAREFUL WITH THIS OPTION
AS FILES CAN BE DELETED FROM YOUR SYSTEM.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Checks if files have the specified file type.
Returns filenames of files that mismatch the file type specified
with an option to delete them.
}
\examples{
\dontrun{
util_check_file_type(
  dir = getwd(),
  fn = c(
    "file_01.pdf",
    "file_02.pdf",
    "file_03.pdf",
    "file_04.pdf",
    "file_05.pdf"
  ),
  file_type = "PDF document",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
