% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_clean_dir.R
\name{util_clean_dir}
\alias{util_clean_dir}
\title{Clean Contents of a Directory}
\usage{
util_clean_dir(dir, create_dir = TRUE, recursive = TRUE, force = TRUE)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{create_dir}{Logical.
If \code{TRUE}, creates an empty \code{dir}.
If \code{FALSE}, \code{dir} and all its contents are deleted.}

\item{recursive}{logical.  Should directories be deleted recursively?}

\item{force}{logical.  Should permissions be changed (if possible) to
    allow the file or directory to be removed?}
}
\description{
Clean Contents of a Directory
}
\examples{
\dontrun{
util_clean_dir(dir = MY_DIRECTORY)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
