% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_clean_tempdir.R
\name{util_clean_tempdir}
\alias{util_clean_tempdir}
\title{Clean Contents of \code{tempdir}}
\usage{
util_clean_tempdir()
}
\description{
Cleans contents of \code{tempdir}.
}
\details{
From
\href{https://stackoverflow.com/users/5783745/stevec}{stevec's}
answer in
\href{https://stackoverflow.com/questions/45894133/deleting-tmp-files}{Stackoverflow}
}
\examples{
\dontrun{
util_clean_tempdir()
}
}
