% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compress.R
\name{util_compress}
\alias{util_compress}
\title{Compress saved R Objects}
\usage{
util_compress(
  dir = getwd(),
  recursive = FALSE,
  format = "Rds",
  pattern = "^filename.*",
  compress = "xz",
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory which contains target files.}

\item{recursive}{Logical.
Search files recursively under \code{dir}.}

\item{format}{Character string.
\code{"Rda"}, \code{"rda"}, \code{"RDA"}, \code{"RData"}, \code{"Rdata"}, \code{"rdata"}, \code{"RDATA"} for \verb{R Data Format}.
\code{"Rds"}, \code{"rds"}, \code{"RDS"} for serialized \code{R} object.}

\item{pattern}{Character string.
Regular expression.
Pattern of file names.
\code{format} is appended as an extension.
For example, if \code{pattern = "^filename.*"},
and \code{format = "Rds"},
the pattern used to load files will be
\code{"^filename.*\\\\.Rds$"}.}

\item{compress}{Character string.
Compression type
(\code{gzip}, \code{bzip2} or \code{xz}).}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Compress saved R Objects
}
\author{
Ivan Jacob Agaloos Pesigan
}
