% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distro.R
\name{util_distro}
\alias{util_distro}
\title{Get Linux Distribution.}
\usage{
util_distro()
}
\value{
Returns the \code{NAME} field in \verb{/etc/os-release} or \verb{/usr/lib/os-release}
in lower case if \code{util_os() == "linux"}.
If the operating system is NOT Linux,
returns the result of \code{\link[=util_os]{util_os()}} .
}
\description{
Utility to detect the Linux distribution
of the current \code{R} session.
}
\examples{
util_distro()
}
\author{
Ivan Jacob Agaloos Pesigan
}
