% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn2columns.R
\name{util_fn2columns}
\alias{util_fn2columns}
\title{Filenames to Columns.}
\usage{
util_fn2columns(df, fn = "fn", split = "_|\\\\.", colnames = NULL)
}
\arguments{
\item{df}{R object.
Data frame.}

\item{fn}{Character string.
Column name of filenames.
Defaults to \code{fn = "fn"}.}

\item{split}{Character string.
Regular expression used to split \code{fn}.
Defaults to \code{split = "_|\\\\."}
to split using \code{"_"} or \code{"."}.}

\item{colnames}{Character vector.
Column names to use to name the new columns produced.
This should have the same length as the columns produced.}
}
\description{
Splits filenames column to create new columns.
See \code{\link[=util_bind]{util_bind()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
