% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_git_config.R
\name{util_git_config}
\alias{util_git_config}
\title{Git config}
\usage{
util_git_config(
  name = "Ivan Jacob Agaloos Pesigan",
  email = "learn.jeksterslab@gmail.com",
  editor = "vim",
  ignore = c("*~", ".*.swp", ".DS_Store", ".~lock.*"),
  msg = paste0("Subject line (try to keep under 50 characters)", "\\n\\n",
    "Multi-line description of commit,", "\\n", "feel free to be detailed.", "\\n\\n",
    "[Ticket: X]"),
  global = TRUE
)
}
\arguments{
\item{name}{Character string.
\code{Git} user.name`.}

\item{email}{Character string.
\code{Git} user.email`.}

\item{editor}{Character string.
\verb{Git core.editor}.}

\item{ignore}{Character vector.
Patterns of file names and paths to ignore.}

\item{msg}{Character string.
Git \code{commit.template}.}

\item{global}{Logical.
If \code{TRUE},
uses the \code{--global} option.
If \code{FALSE},
uses the \code{--local} option.}
}
\description{
Git config
}
