% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_subdir.R
\name{util_make_subdir}
\alias{util_make_subdir}
\title{Make Subdirectories}
\usage{
util_make_subdir(dir = NULL, subdir = NULL)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{subdir}{Character string or vector.
A single subdirectory of a vector of subdirectories.
Subdirectory or subdirectories.}
}
\description{
Makes subdirectories recursively.
}
\details{
For example,
if \code{dir = "/home/jek"}
and
\code{subdir = "subdir"}
\code{\link[=util_make_subdir]{util_make_subdir()}}
creates
\verb{/home/jek/subdir}.
If \code{dir = "/home/jek"}
and
\code{subdir = c("subdir1", "subdir2")}
\code{\link[=util_make_subdir]{util_make_subdir()}}
creates
\verb{/home/jek/subdir1/subdir2}.
If \code{dir} and \code{subdir} are not provided
\code{\link[=util_make_subdir]{util_make_subdir()}}
creates
a subdirectory with a random name
(see \code{\link[=util_rand_str]{util_rand_str()}})
in the home directory.
}
\author{
Ivan Jacob Agaloos Pesigan
}
