% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_os.R
\name{util_os}
\alias{util_os}
\title{Get Operating System.}
\usage{
util_os()
}
\value{
Return any of the following:
\itemize{
\item \code{linux} for GNU/Linux operating system
\item \code{osx} for Macintosh operating system
\item \code{windows} for Windows operating system
\item mystery machine
}
}
\description{
Utility to detect the operating system of the current \code{R} session.
Based on
\url{http://conjugateprior.org/2015/06/identifying-the-os-from-r/}.
}
\examples{
util_os()
}
