% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pkgload.R
\name{util_pkgload}
\alias{util_pkgload}
\title{Load Packages.}
\usage{
util_pkgload(
  pkg,
  lib.loc = .libPaths()[1],
  repos = "https://cran.rstudio.org",
  dependencies = TRUE,
  type = "source",
  update = FALSE
)
}
\arguments{
\item{pkg}{Character vector of packages.}

\item{lib.loc}{character vector describing the location of R
    library trees to search through (and update packages therein), or
    \code{NULL} for all known trees (see \code{\link{.libPaths}}).}

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, e.g., the URL of a CRAN mirror such as
    \code{"https://cloud.r-project.org"}.
  }

\item{dependencies}{logical indicating whether to also install
    uninstalled packages which these packages depend on/link
    to/import/suggest (and so on recursively).
    Not used if \code{repos = NULL}.
    Can also be a character vector, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.

    Only supported if \code{lib} is of length one (or missing),
    so it is unambiguous where to install the dependent packages.  If
    this is not the case it is ignored, with a warning.

    The default, \code{NA}, means
    \code{c("Depends", "Imports", "LinkingTo")}.

    \code{TRUE} means to use
    \code{c("Depends", "Imports", "LinkingTo", "Suggests")} for
    \code{pkgs} and
    \code{c("Depends", "Imports", "LinkingTo")} for added dependencies:
    this installs all the packages needed to run \code{pkgs}, their
    examples, tests and vignettes (if the package author specified them
    correctly).

    In all of these, \code{"LinkingTo"} is omitted for binary packages.
  }

\item{type}{character, indicating the type of package to download and
    install.  See \code{\link[utils]{install.packages}}.
  }

\item{update}{Logical.
Update installed packages.}
}
\description{
Checks if a character vector of \code{R} packages are installed,
installs them if they are not currently installed,
and loads them in the namespace.
}
\examples{
pkg <- list(
  "testthat",
  "devtools",
  "rmarkdown"
)
pkg_load(pkg = pkg)
}
\author{
Ivan Jacob Agaloos Pesigan
}
