% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_render.R
\name{util_render}
\alias{util_render}
\title{Render R Scripts and R Markdown Files.}
\usage{
util_render(dir = getwd(), files = NULL, par = TRUE, ncores = NULL)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{files}{Character vector.
Vector of files to render.
If \code{NULL},
the function will recursively render
all R scripts \code{.R} and
R Markdown files \code{.Rmd}
in \code{dir}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Render R Scripts and R Markdown Files.
}
\examples{
\dontrun{
util_render()
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
