% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_renviron.R
\name{util_renviron}
\alias{util_renviron}
\title{Create \code{.Renviron} File}
\usage{
util_renviron(
  dir = Sys.getenv("HOME"),
  overwrite = FALSE,
  GITHUB_PAT = NULL,
  libpath = NULL
)
}
\arguments{
\item{dir}{Character string.
\code{.Renviron} directory}

\item{overwrite}{Logical.
Overwrite existing \code{.Renviron} file in \code{dir}.}

\item{GITHUB_PAT}{Character string.
Optional argument.
Github Personal Access Token.}

\item{libpath}{Character string.
Library path.
If unspecified, defaults to
\code{{HOME}/R/{PLATFORM}-library/{R.VERSION}}.}
}
\description{
Creates a \code{.Renviron} file in \code{dir}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
