% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rprofile.R
\name{util_rprofile}
\alias{util_rprofile}
\title{Create \code{.Rprofile} File}
\usage{
util_rprofile(dir = Sys.getenv("HOME"), overwrite = FALSE)
}
\arguments{
\item{dir}{Character string.
\code{.Rprofile} directory}

\item{overwrite}{Logical.
Overwrite existing \code{.Rprofile} file in \code{dir}.}
}
\description{
Creates a \code{.Rprofile} file in \code{dir}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
