% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_search.R
\name{util_search_ext}
\alias{util_search_ext}
\title{Search for Files using File Extension}
\usage{
util_search_ext(
  dir = getwd(),
  ext = "csv",
  all.files = FALSE,
  full.names = TRUE,
  recursive = FALSE,
  ignore.case = TRUE,
  no.. = FALSE
)
}
\arguments{
\item{dir}{Character string.
Directory to search.}

\item{ext}{Character string.
File extension.}

\item{all.files}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned (following Unix-style visibility,
    that is files whose name does not start with a dot).  If \code{TRUE},
    all file names will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive?}

\item{no..}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
}
\description{
Searches for files in \code{dir}
using file extension.
}
\examples{
# Search for files in the working directory with the csv extension
util_search_ext(
  dir = getwd(),
  ext = "csv"
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
