% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_search.R
\name{util_search_pattern}
\alias{util_search_pattern}
\title{Search for Files using Regular Expression}
\usage{
util_search_pattern(
  dir = getwd(),
  pattern = "^file.*\\\\.csv$",
  all.files = FALSE,
  full.names = TRUE,
  recursive = FALSE,
  ignore.case = TRUE,
  no.. = FALSE
)
}
\arguments{
\item{dir}{Character string.
Directory to search.}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{all.files}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned (following Unix-style visibility,
    that is files whose name does not start with a dot).  If \code{TRUE},
    all file names will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive?}

\item{no..}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
}
\description{
Searches for files in \code{dir}
using file name regular expression pattern.
}
\examples{
# Search for files in the working directory
# with filenames starting with `file`
# and ending with `.csv`.
util_search_pattern(
  dir = getwd(),
  pattern = "^file.*\\\\.csv$"
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
