% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_search.R
\name{util_search_r}
\alias{util_search_r}
\title{Search for R and R Markdown Files}
\usage{
util_search_r(
  dir = getwd(),
  rscript = TRUE,
  rmd = TRUE,
  all.files = FALSE,
  full.names = TRUE,
  recursive = FALSE,
  ignore.case = TRUE,
  no.. = FALSE
)
}
\arguments{
\item{dir}{Character string.
Directory to search.}

\item{rscript}{Logical.
Search for \code{R} scripts (\verb{*.R} or \verb{*.r}).}

\item{rmd}{Logical.
Search for \verb{R Markdown} files (\verb{*.Rmd} or \verb{*.rmd}).}

\item{all.files}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned (following Unix-style visibility,
    that is files whose name does not start with a dot).  If \code{TRUE},
    all file names will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive?}

\item{no..}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
}
\description{
Searches for \code{R} and \verb{R Markdown} files in \code{dir}.
}
\examples{
util_search_r(dir = getwd())
}
\author{
Ivan Jacob Agaloos Pesigan
}
