% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_spin.R
\name{util_spin}
\alias{util_spin}
\title{Spin \code{R} Scripts into \code{R} Markdown Files.}
\usage{
util_spin(
  dir = getwd(),
  recursive = FALSE,
  files = NULL,
  knit = FALSE,
  par = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{dir}{Character string.
Directory.
Ignored if \code{file} is NOT NULL.}

\item{recursive}{Logical.
If \code{TRUE},
recursively spin all \code{R} scripts (\code{.R}, \code{.r})
in \code{dir}.
Ignored if \code{file} is NOT NULL.}

\item{files}{Character vector.
Vector of files to spin.}

\item{knit}{Logical.
If \code{TRUE}, compiles the document after conversion.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{...}{Arguments to pass to \code{\link[knitr:spin]{knitr::spin()}}.}
}
\description{
Spin \code{R} Scripts into \code{R} Markdown Files.
}
\examples{
\dontrun{
util_spin(
  dir = getwd(),
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
