% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_txt2file.R
\name{util_txt2file}
\alias{util_txt2file}
\title{Text to File.}
\usage{
util_txt2file(text, dir, fn, msg = NULL)
}
\arguments{
\item{text}{A character vector}

\item{dir}{Character string.
Directory.}

\item{fn}{Character string.
Filename.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Writes lines of character strings to file.
}
\examples{
text <- paste0(
  "Lorem ipsum dolor sit amet,",
  "consectetur adipiscing elit,",
  "sed do eiusmod tempor incididunt",
  "ut labore et dolore magna aliqua.",
  "Ut enim ad minim veniam,",
  "quis nostrud exercitation ullamco laboris nisi",
  "ut aliquip ex ea commodo consequat.",
  "Duis aute irure dolor in reprehenderit in voluptate velit",
  "esse cillum dolore eu fugiat nulla pariatur.",
  "Excepteur sint occaecat cupidatat non proident,",
  "sunt in culpa qui officia deserunt mollit anim id est laborum."
)
util_txt2file(
  text = text,
  dir = getwd(),
  fn = "Lipsum.txt",
  msg = "Output file:"
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
