% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_url_exists.R
\name{util_url_exists}
\alias{util_url_exists}
\title{Check URL}
\usage{
util_url_exists(con)
}
\arguments{
\item{con}{URL.}
}
\description{
Checks is website exists.
}
\author{
Ivan Jacob Agaloos Pesigan
}
