% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_user_lib.R
\name{util_user_lib}
\alias{util_user_lib}
\title{Set User Library}
\usage{
util_user_lib(libpath = NULL)
}
\arguments{
\item{libpath}{Character string.
Library path.
If unspecified, defaults to
\code{{HOME}/R/{PLATFORM}-library/{R.VERSION}}.}
}
\description{
Set User Library
}
\author{
Ivan Jacob Agaloos Pesigan
}
