% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_valid_email.R
\name{util_valid_email}
\alias{util_valid_email}
\title{Validate Email Address}
\usage{
util_valid_email(x)
}
\arguments{
\item{x}{Character string.
Email address to validate.}
}
\description{
Validate Email Address
}
\author{
Ivan Jacob Agaloos Pesigan
}
