% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_vim.R
\name{util_vim}
\alias{util_vim}
\title{Create \code{.vimrc} File and Install Plugins}
\usage{
util_vim(dir = Sys.getenv("HOME"), overwrite = FALSE, plugins = FALSE)
}
\arguments{
\item{dir}{Character string.
\code{.vimrc} directory}

\item{overwrite}{Logical.
Overwrite existing \code{.vimrc} file in \code{dir}.}

\item{plugins}{Logical.
Install \code{vim} plugins.}
}
\description{
Creates a \code{.vimrc} file in \code{dir}
and install plugins.
}
\author{
Ivan Jacob Agaloos Pesigan
}
