% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wget.R
\name{util_wget}
\alias{util_wget}
\title{wget}
\usage{
util_wget(dir = getwd(), link, args = "-nc", par = TRUE, ncores = NULL)
}
\arguments{
\item{dir}{Character string.
Directory where the downloaded file is saved.}

\item{link}{Character vector.
URLs.}

\item{args}{Character string.
Arguments to pass to \code{wget}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
wget
}
\examples{
\dontrun{
link <- c(
  "https://raw.githubusercontent.com/jeksterslabds/jeksterslabRutils/master/R/util_bind.R",
  "https://raw.githubusercontent.com/jeksterslabds/jeksterslabRutils/master/R/util_cat_sys.R",
)
util_wget(
  dir = getwd(),
  link = link,
  args = "-nc",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
