% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_xml2list.R
\name{util_xml2list}
\alias{util_xml2list}
\title{XML/HTML to list}
\usage{
util_xml2list(tags, con)
}
\arguments{
\item{tags}{Character vector.
XML or HTML tags.}

\item{con}{Connection.
Path or connection to XML/HTML file.}
}
\description{
XML/HTML to list
}
\author{
Ivan Jacob Agaloos Pesigan
}
