% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_clean_tempdir.R
\name{util_clean_tempdir}
\alias{util_clean_tempdir}
\title{Clean \code{tempdir}.}
\usage{
util_clean_tempdir()
}
\description{
Clean \code{tempdir}.
}
\examples{
\dontrun{
util_clean_tempdir()
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
