% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_numbers.R
\name{util_get_numbers}
\alias{util_get_numbers}
\title{Extract Numbers from String.}
\usage{
util_get_numbers(x)
}
\arguments{
\item{x}{String.}
}
\description{
Extract Numbers from String.
}
\examples{
x <- "fkj;lakjfajkfak73875kllkj75487"
util_get_numbers(x)
}
\author{
Ivan Jacob Agaloos Pesigan
}
