% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list2vector.R
\name{util_list2vector}
\alias{util_list2vector}
\title{Extract items from a list to a vector.}
\usage{
util_list2vector(fields, index)
}
\arguments{
\item{fields}{Character vector.
Vector of names in the list.}

\item{index}{List.}
}
\description{
Extract items from a list to a vector.
}
\author{
Ivan Jacob Agaloos Pesigan
}
