% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list2vector.R
\name{util_list2vector}
\alias{util_list2vector}
\title{List to Vector.}
\usage{
util_list2vector(fields, index)
}
\arguments{
\item{fields}{Character vector.
Vector of names in the list.}

\item{index}{List.}
}
\description{
Extracts items from a list and returns a vector.
Note that this only works with lists
that follows the structure in the example below.
}
\examples{
index <- list(
  firstname = "Ivan Jacob",
  lastname = "Pesigan",
  province = "BC",
  number = 27
)
util_list2vector(
  fields = c(
    "firstname",
    "lastname",
    "province",
    "number"
  ),
  index = index
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
