% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_render.R
\name{util_render}
\alias{util_render}
\title{Render \code{R} Scripts and \code{R} Markdown Files.}
\usage{
util_render(
  dir = getwd(),
  recursive = FALSE,
  files = NULL,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory.
Used if \code{recursive = TRUE}.}

\item{recursive}{Logical.
If \code{TRUE},
recursively render all \code{R} scripts (\code{.R}, \code{.r}) and
R Markdown files (\code{.Rmd}, \code{.rmd})
in \code{dir}.}

\item{files}{Character vector.
Vector of files to render.
Used if \code{recursive} is \code{FALSE}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Render \code{R} Scripts and \code{R} Markdown Files.
}
\examples{
\dontrun{
util_render(
  dir = getwd(),
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
