% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shiny.R
\name{util_shiny}
\alias{util_shiny}
\title{View Data Using a Shiny App}
\usage{
util_shiny(app, package)
}
\arguments{
\item{app}{Character string.
Name of the Shiny App.}
}
\description{
inst/shiny/app
}
\author{
Ivan Jacob Agaloos Pesigan
}
