% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_style.R
\name{util_style}
\alias{util_style}
\title{Style \code{R} and \code{R} Markdown Files.}
\usage{
util_style(
  dir = getwd(),
  recursive = FALSE,
  files = NULL,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{recursive}{Logical.
If \code{TRUE},
recursively style all \code{R} scripts (\code{.R}, \code{.r}) and
R Markdown files (\code{.Rmd}, \code{.rmd})
in \code{dir}.}

\item{files}{Character vector.
Vector of files to style.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\description{
Styles all \code{R} scripts and \code{R} Markdown files
in a given directory.
}
\examples{
\dontrun{
util_style(
  dir = getwd(),
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
