% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_url_exists.R
\name{util_url_exists}
\alias{util_url_exists}
\title{Check URL}
\usage{
util_url_exists(con)
}
\arguments{
\item{con}{URL.}
}
\description{
Checks if website exists.
}
\examples{
util_url_exists("https://www.google.com")
}
\author{
Ivan Jacob Agaloos Pesigan
}
