% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_user_lib.R
\name{util_user_lib}
\alias{util_user_lib}
\title{Set User Library}
\usage{
util_user_lib(libpath = NULL, dir = Sys.getenv("HOME"))
}
\arguments{
\item{libpath}{Character string.
Library path.
If unspecified, defaults to
\code{{HOME}/R/{PLATFORM}-library/{R.VERSION}}.}

\item{dir}{Character string.
\code{.Renviron} directory}
}
\description{
Sets user library using \code{.libPaths}
and \code{R_LIBS_USER} in \code{.Renviron}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
