% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_xml2list.R
\name{util_xml2list}
\alias{util_xml2list}
\title{\code{XML}/\code{HTML} to List.}
\usage{
util_xml2list(tags, con, par = TRUE, ncores = NULL)
}
\arguments{
\item{tags}{Character vector.
\code{XML} or \code{HTML} tags.}

\item{con}{Connection.
Path or connection to \code{XML}/\code{HTML} file.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\description{
\code{XML}/\code{HTML} to List.
}
\examples{
xml <- paste0(
  "<TITLE>",
  "Romanza",
  "</TITLE>",
  "<ARTIST>",
  "Andrea Bocelli",
  "</ARTIST>",
  "<COUNTRY>",
  "EU",
  "</COUNTRY>",
  "<COMPANY>",
  "Polydor",
  "</COMPANY>",
  "<PRICE>",
  "10.80",
  "</PRICE>",
  "<YEAR>",
  "1996",
  "</YEAR>"
)
tmp <- tempfile()
writeLines(
  text = xml,
  con = tmp
)
tags <- c(
  "TITLE",
  "ARTIST",
  "COUNTRY",
  "COMPANY",
  "PRICE",
  "YEAR"
)
util_xml2list(
  tags = tags,
  con = tmp,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
